Fixed tests, also renamed some test classes as unity would confuse them when trying to go to the source code of those tests
