+++ unity-renderer/Assets/DCLPlugins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e9172fb83854a82bf45d1dd650af779
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
