+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/AssetBundles/AB_GameObject/Tests/ABGoBlockedAndMasterPromisesShould.cs
@@ -8,7 +8,7 @@ using UnityEngine.TestTools;
 
 namespace AssetPromiseKeeper_AssetBundle_GameObject_Tests
 {
-    public class BlockedAndMasterPromisesShould : TestsBase_APK<AssetPromiseKeeper_AB_GameObject,
+    public class ABGoBlockedAndMasterPromisesShould : TestsBase_APK<AssetPromiseKeeper_AB_GameObject,
                                                 AssetPromise_AB_GameObject,
                                                 Asset_AB_GameObject,
                                                 AssetLibrary_AB_GameObject>
