+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/AudioClip/Tests/ACBlockedAndMasterPromisesShould.cs
@@ -1,44 +1,35 @@
+﻿using System.Collections;
 using AssetPromiseKeeper_Tests;
 using DCL;
-using DCL.Helpers;
-using System.Collections;
-using System.Text.RegularExpressions;
-using UnityEngine;
 using UnityEngine.Assertions;
 using UnityEngine.TestTools;
 
-namespace AssetPromiseKeeper_AssetBundle_Tests
+namespace AssetPromiseKeeper_AudioClip_Tests
 {
-    public class BlockedAndMasterPromisesShould :
-        TestsBase_APK<AssetPromiseKeeper_AB, AssetPromise_AB, Asset_AB, AssetLibrary_AB>
+    public class ACBlockedAndMasterPromisesShould : TestsBase_APK<AssetPromiseKeeper_AudioClip,
+        AssetPromise_AudioClip,
+        Asset_AudioClip,
+        AssetLibrary_RefCounted<Asset_AudioClip>>
     {
-        protected AssetPromise_AB CreatePromise(string hash = null)
-        {
-            string contentUrl = TestAssetsUtils.GetPath() + "/AssetBundles/";
-            hash = hash ?? "QmYACL8SnbXEonXQeRHdWYbfm8vxvaFAWnsLHUaDG4ABp5";
-            var prom = new AssetPromise_AB(contentUrl, hash);
-            return prom;
-        }
-
         [UnityTest]
         public IEnumerator FailCorrectlyWhenGivenWrongURL()
         {
-            string url = "non_existing_url.glb";
+            var provider = new ContentProvider_Dummy();
 
-            AssetPromise_AB prom = CreatePromise(url);
-            Asset_AB asset = null;
+            var prom = new AssetPromise_AudioClip("123325", provider);
+            Asset_AudioClip asset = null;
             bool failEventCalled1 = false;
             prom.OnSuccessEvent += (x) => { asset = x; };
             prom.OnFailEvent += (x, error) => { failEventCalled1 = true; };
 
-            AssetPromise_AB prom2 = CreatePromise(url);
-            Asset_AB asset2 = null;
+            var prom2 = new AssetPromise_AudioClip("43254378", provider);
+            Asset_AudioClip asset2 = null;
             bool failEventCalled2 = false;
             prom2.OnSuccessEvent += (x) => { asset2 = x; };
             prom2.OnFailEvent += (x, error) => { failEventCalled2 = true; };
 
-            AssetPromise_AB prom3 = CreatePromise(url);
-            Asset_AB asset3 = null;
+            var prom3 = new AssetPromise_AudioClip("09898765", provider);
+            Asset_AudioClip asset3 = null;
             bool failEventCalled3 = false;
             prom3.OnSuccessEvent += (x) => { asset3 = x; };
             prom3.OnFailEvent += (x, error) => { failEventCalled3 = true; };
@@ -46,6 +37,8 @@ namespace AssetPromiseKeeper_AssetBundle_Tests
             keeper.Keep(prom);
             keeper.Keep(prom2);
             keeper.Keep(prom3);
+
+            Assert.AreEqual(3, keeper.waitingPromisesCount);
 
             yield return prom;
             yield return prom2;
